Multiprocessing in Mac appears to have changes
Multiprocessing package is changed from python 3.8 for MacOs. Attempting
to force a change to work across all OS.

+++ tools/cpp/KML/include/statistics/summary/IStreamingStatistic.tcc
@@ -8,7 +8,7 @@ namespace KML
     namespace Statistics
     {
         template<typename T>
-        IStreamingStatistic<T>::IStreamingStatistic(const uint64_t& windowSize)
+        IStreamingStatistic<T>::IStreamingStatistic(const size_t& windowSize)
             : IStreaming<T>(windowSize)
         {
             // Do nothing.
